line-bot-python/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Line Bot SDK
line-bot-sdk==3.8.0

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AI/ML
google-genai

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging
loguru==0.7.2

# Optional: For better performance
orjson==3.9.12
